On the iPhone, if JS runs for longer than 5 seconds, then Safari thinks it's runaway code and immediately halts the current function and passes "undefined" back to its caller.
For Galatea, the base64 decoding of the relatively massive 442k zblorb was taking much longer than 5 seconds, so I made a function that "wimped out" after decoding 50k increments.  Now Galatea works, it just takes a really long time to load.  IIRC, the browser has a built-in mime decoding function that should be a lot faster than the JS implementation being used now, I just couldn't figure out how to get it to work the way I wanted earlier... Another alternative is to use sqlite to cache some of this stuff, which I think the version of Safari on the iPhone supports.