Some Unicode fixes for issues 26, 56 and 62.
All Unicode characters should now be downcased correctly before being encoded to be compared with dictionary words.

Conversions from Unicode to ZSCII have been made faster with a reverse lookup table.

Printing Unicode characters to Z-machine memory using stream 3 should now convert them to ZSCII before storing them.

There is a new function to convert a character code from Unicode to ZSCII:
_ascii_code_to_zscii_code(ascii_code)

Unicode is refered to as ASCII in function names in the rest of the code, so that convention was kept here.

It takes a Unicode code as an argument, and returns the corresponding ZSCII code. If that character doesn't have a ZSCII code, the ZSCII/ASCII code for '*' is returned instead.